Fix test gaps in Encoder/MapcodeCodec
A few things going on in this commit, all part of the quest for 100% coverage.

First of all, we were double-checking arguments in MapcodeCodec; this is
pointless and leads to uncoverable code in Encoder. Now all argument checks are
done in Encoder.

We still need to check responses on the way out of MapcodeCodec since it calls
the Encoder API in various ways, and needs to verify the results in a way
Encoder cannot.